jedi/requirements.txt
confluent-kafka==2.3.0
minio==7.2.5
pyarrow==16.1.0
pyspark==3.5.1
pyyaml==6.0.1
delta-spark == 3.2.0